requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bech32==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cosmpy==0.9.2
ecdsa==0.19.1
frozenlist==1.5.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.2.0
propcache==0.3.0
protobuf==4.25.6
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame-ce==2.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyTMX==3.32
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
typing_extensions==4.12.2
uagents==0.21.0
uagents-ai-engine==0.9.0
urllib3==2.3.0
uvicorn==0.30.6
wheel==0.45.1
yarl==1.18.3
